Fix(core): Remove duplicate dashboard initialization
Removes the redundant code block that initialized and started the dashboard server twice. This prevents potential race conditions and resource conflicts.

Refactor(utils): Centralize utility functions

- Moves the `formatNumber` function from `ytSubRole.js` to `stringUtils.js` to promote code reuse.
- Updates the `ping` command to use the existing `formatUptime` utility function instead of a local implementation.
