+++ src/features/youtube-subscriber-roles/commands/ytSubRole.js
@@ -10,6 +10,7 @@ const {
 const { google } = require("googleapis");
 const axios = require("axios");
 const { handleError } = require("../../../utils/errorHandler");
+const { formatNumber } = require("../../../utils/stringUtils");
 const YTSubRoleConfig = require("../database/ytSubRoleConfig");
 const TempOAuth2Storage = require("../database/tempOAuth2Storage");
 
@@ -406,17 +407,3 @@ module.exports = {
     },
 };
 
-// Helper function to format numbers
-function formatNumber(num) {
-    if (!num) return "0";
-    if (num >= 1000000000) {
-        return (num / 1000000000).toFixed(1) + "B";
-    }
-    if (num >= 1000000) {
-        return (num / 1000000).toFixed(1) + "M";
-    }
-    if (num >= 1000) {
-        return (num / 1000).toFixed(1) + "K";
-    }
-    return num.toString();
-}
